requirements.txt
streamlit
pandas
plotly
openpyxl
openai
requests
pytest
python-dotenv
